+++ tests/run.clj
@@ -89,6 +89,7 @@
 (require "tests/instant_template_test")
 (require "tests/data_test")
 (require "tests/data_readers_test")
+(require "tests/reader_conditional_test")
 (require "tests/spec_test")
 (require "tests/reducers_test")
 (require "tests/capability_metadata_test")